adws/pyproject.toml
[project]
name = "adws"
version = "0.1.0"
description = "Autonomous Developer Workflow System — Trinity Protocol orchestration for ThemeGPT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "openai>=2.24.0",
    "google-genai>=1.65.0",
    "fastapi>=0.133.0",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.0",
    "rich>=14.0.0",
    "typer>=0.24.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".."]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["adw_modules", "scripts"]
